pyproject.toml
[project]
name = "sefaz-ce"
version = "0.1.0"
description = "Automação de consulta DAE no portal SEFAZ-CE"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
sefaz-ce-gui = "sefaz_ce.gui_app:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sefaz_ce*"]

[tool.uv]
dev-dependencies = []
